requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.7
dataclasses-json==0.6.7
edge-tts==7.0.2
ffmpeg-python==0.2.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain-community
langchain-core
langchain-groq
langsmith
nest-asyncio
numpy
orjson
packaging
pandas
pexels-api==1.0.1
pillow
pip
protobuf
pyarrow
pydantic
pydantic-core
pydantic-settings
pydeck
python-dotenv
PyYAML
referencing
requests
requests-toolbelt
rpds-py
smmap
sniffio
SQLAlchemy
streamlit
tabulate
tenacity
toml
tornado
tqdm
typing-inspect
typing-inspection
typing_extensions
tzdata
urllib3
yarl
zstandard